Add files via upload
Simplex_algorithm_final.ipynb is a jupyter notebook (code written in Python 3 kernel) which takes data (cost coefficient, decision variables, b vector, A matrix of the simplex method) and solves it. Here all constraints must be of the form <= and all values of b , A must be positive. This solves for maximization problem. simplex_data.csv has a column called A where the value of matrix A must be entered row wise i.e 1 st cell of A column is a11, second cell contains value of a12 and so on.